+++ akka-http-cors/src/main/scala/ch/megard/akka/http/cors/javadsl/CorsRejection.scala
@@ -16,7 +16,13 @@ object CorsRejection {
   /**
     * Signals the cause of the failed CORS request.
     */
-  trait Cause
+  trait Cause {
+    /**
+      * Description of this Cause in a human-readable format.
+      * Can be used for debugging or custom Rejection handlers.
+      */
+    def description: String
+  }
 
   /**
     * Signals the CORS request was malformed.